Reshape cycles to aliquots all at once
This is another outrageously large commit. Basically deletes everything
from the previous commit and rewrites it so that I identify the blocks
and aliquots all within cycle_metadata and reshape it into
aliquot_deltas in one step, i.e. without the intermediary block_deltas.
This is the way I should have done it from the beginning tbh. It also
includes the functionality from the previous commit. reshapeCycles can
still output PIS deltas and metadata and a cell array of all the blocks
and or aliquots.