Add missing time intervals to sample type timeline
Add patch shading to indicate the time intervals that are not covered
by any of the spreadsheet files that I have. This way I can compare to
the sample type timeline to see if there were any samples run during
the missing intervals or if the MS was idle and there is no excel file
containing metadata that I need to track down.