Make stats output from calculateChemSlope same format as output calculatePisValues
Rather than output the stats in a structure array with one entry for
each chem slope experiment, the function now mimics the output from
calculatePisValues. The stats are stored in a structure of arrays of
the same size as aliquot deltas. Some of the entries are cell arrays to
allow for a different number of x or y data in different CS experiments