Calculate new LJA Values after each refocus/new filament/std can swap
Modifies calculateLjaValues.m to split the different LJA aliquots into
groups based on the changes to the mass spectrometer (i.e. a new
filament, refocussing of the source, or change in the standard can).
Calculate one mean for each delta value for each correction period and
also fit a linear slope to the data and measure its significance to
allow me to check for temporal variation in LJA values.