Write function to make LJA Corr using mean of aliquots or linear trend
Write a function that makes the LJA correction on raw delta values. The
means of all LJA aliquots measured under identical MS conditions are
used, unless there is a trend in the aliquots. In this case, the linear
trend of the LJA aliqouts is extrapolated to give the LJA values for
all aliquots.