Change method of identifying Chem Slope Aliquots
In order to extend the record back to 2013, I have to be more flexible
in the method I use to identify a Chem Slope. Some CS aliquots do not
contain "CS" in their ID1. This commit uses regular expressions to find
aliquots with (N or O) and (a replicate number), that may or may not
also contain "15", "18", and "CS", and label them as CS aliquots.
This works fine back to 2013 but an even more flexible way might just
be to look at the elemental ratios that should be abnormally large in
magnitude for the CS aliquots. Something to think about.