Update spcDataPull to use makeMasterSheet
Previously, spcDataPull needed several other scripts to have been run
as it used the variables that they created in the workspace without
generating them itself. Now it is a self-contained script that can be
run alone. It uses makeMasterSheet to generate the data it needs.

This commit also condenses a lot of redundant code into for loops,
massively simplifying the code.