Create and write makeIceCoreDataset.m
Write a function that assembles an Ice Core Dataset from a masterSheet
produced by makeMasterSheet. The function searches for an ice core ID
and compiles measurements from this core that have the same bottom
depth into a table of replicates. This function replaces the previous
spcDataPull.m script, and can be used to generate datasets for
different ice cores.