Update PIS correction to pass variables using structures
This commit updates the PIS section of makeMasterSheet and the two
relevant subfunctions, calculatePisValues and makePisCorr, to pass
variables using structures.

N.B. This version of makeMasterSheet will not run as the following
sections of the function have not yet been updated. See the next few
commit messages for these updates.