Add path/filename identifier in new file name in PS renaming script
Adds characters that split the prepended relative path from the
original file name in the new file name generated by the powershell
renaming script.

Additionally, makes a few other small tweaks to the new file name and
checks it doesn't exceed the 260 character length for a file path.
Exits execution if it does.