Modify reshapeCycles to handle samples from pre-automation era
Previously, reshapeCycles identified the different sample aliquots
based on the automation methods that were run at the start of every new
sample. This doesn't work in the pre-automation era.

Instead the method now identifies samples based on a change in the
inferred user-entered file name, which changes for (almost) every new
sample.