Fix calcCycles.m and makeRawDataset.m to work for both ISODAT 2 and 3
Insert an if statement so that the section of makeRawDataset that deals
with ISODAT 2 checks doesn't trip up the function if only ISODAT 3
files are run.

Also merge the 'h2o' and 'co2 check' fields into one 'checks' field and
give the CO2 B and CO2-nonB checks the same variables and names.