Update makeIceCoreDataset.m to be more flexible
The makeIceCoreDataset.m function now accepts different inputs that
make it easier to work with the many different metadata formats that
exist for different ice cores. Rather than accepting a regular
expression to match to, this step is performed outside of the function
and the function is passed a logical vector that identifies the ice
core aliquots. The bottom depths are also determined outside of the
function using the ID1 or file name and are passed to the function as a
third input rather than try to build all of this capability into
makeIceCoreDataset.m itself.