Update makeIceCoreDataset.m to handle depth more explicitly
The makeIceCoreDataset.m no longer assumes that the depths it is given
are bottom depths. Instead it just uses DEPTH as the variable name.
This commit also updates the individual ice core plotting scripts to
handle their depth variable explicitly, renaming the variable in
iceCore.metadata to either topDepth or bottomDepth as appropriate.